requirements.txt
alembic==1.5.7
boto3==1.17.77
botocore==1.20.77
Flask==1.1.2
Flask_Login==0.5.0
Flask_Mail==0.9.1
Flask_Migrate==2.7.0
Flask_RESTful==0.3.9
flask_restplus==0.13.0
Flask_SQLAlchemy==2.4.4
Flask_WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.4
minio==7.0.3
num2words==0.5.10
pdfkit==0.6.1
python-decouple==3.6
requests==2.25.1
SQLAlchemy==1.3.23
SQLAlchemy_serializer==1.4.1
Werkzeug==1.0.1
WTForms==2.3.3
